requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
coverage==7.4.3
cryptography==42.0.5
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==24.1.0
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
mccabe==0.7.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
